merge

+++ src/uet/oop/bomberman/entities/Flame.java
@@ -9,7 +9,8 @@ public class Flame extends Entity implements IAnimation {
     protected int animate = 0;
     protected int frameCount = 0;
 
-    public Flame(int xUnit, int yUnit, Image img) {
+
+    public Flame(double xUnit, double yUnit, Image img) {
         super(xUnit, yUnit, img);
     }
